Create ItinerariesloaderTypes
