+++ apps/graphql/src/dataloaders/Itinerariesloader.js
@@ -7,149 +7,18 @@ import { OptimisticDataloader } from '@kiwicom/graphql-utils';
 
 import itineraryMock from '../mocks/itinerary';
 import fetch from '../services/Fetch';
-
-export type ItinerariesSearchParameters = {|
-  +travelFrom: string,
-  +dateFrom: Date,
-  +dateTo?: Date,
-  +travelTo?: string,
-  +returnDateFrom?: Date,
-  +returnDateTo?: Date,
-  +passengers?: {|
-    +adults?: number,
-    +children?: number,
-    +infants?: number,
-  |},
-|};
-
-export type TripSegment = {|
-  +city: string,
-  +cityCode: string,
-  +localTime: ?Date,
-  +utcTime: ?Date,
-|};
-export type RouteItem = {|
-  +airline: string,
-  +arrival: TripSegment,
-  +departure: TripSegment,
-  +id: string,
-|};
-
-// start - new structure
-
-export type Country = {|
-  +id: string,
-  +name: string,
-  +code: string,
-  +slug: string,
-  +flagURL: string,
-|};
-
-// @TODO rename PriceType
-export type Price = {|
-  +amount: number,
-  +currency: string,
-|};
-
-export type LocationType = {|
-  +id: string,
-  +locationId: string,
-  +name: string,
-  +timezone: string,
-  +country: Country,
-|};
-export type DateType = {|
-  +local: Date,
-  +utc: Date,
-|};
-export type ProviderType = {|
-  +id: string,
-  +name: string,
-|};
-export type VehicleType = {|
-  +type: string,
-  +uniqueNo: string,
-|};
-export type SegmentType = {|
-  +arrivalTime: DateType,
-  +departureTime: DateType,
-  +destination: LocationType,
-  +duration: number,
-  +id: string,
-  +origin: LocationType,
-  +provider: ProviderType,
-  +vehicle: VehicleType,
-|};
-export type SectorType = {|
-  +arrivalTime: DateType,
-  +connections: Array<SegmentType>,
-  +departureTime: DateType,
-  +destination: LocationType,
-  +duration: number,
-  +id: string,
-  +origin: LocationType,
-  +segments: Array<SegmentType>,
-|};
-
-export type newItinerariesStructure = {|
-  id: string,
-  +type: string,
-  +price: Price,
-  +origin: LocationType,
-  +destination: LocationType,
-  +startTime: DateType,
-  +endTime: DateType,
-  +sectors: Array<SectorType>,
-|};
-// end - new structure
-
-export type Itineraries = {|
-  +airlines: Array<string>,
-  +id: string,
-  +flyFrom: string,
-  +flyTo: string,
-  +localDeparture: string,
-  +localArrival: string,
-  +price: Price,
-  +route: Array<RouteItem>,
-  +routes: Array<Array<string>>,
-  ...newItinerariesStructure,
-|};
-
-export type ApiRouteItem = {|
-  +airline?: string,
-  +cityFrom?: string,
-  +cityTo?: string,
-  +flyFrom?: string,
-  +flyTo?: string,
-  +id?: string,
-  +local_arrival?: Date,
-  +utc_arrival?: ?Date,
-  +local_departure?: Date,
-  +utc_departure?: Date,
-|};
-
-type ApiResponse = {|
-  +currency: string,
-  +data: $ReadOnlyArray<{|
-    +id: string,
-    +airlines: Array<string>,
-    +price: number,
-    +flyFrom: string,
-    +flyTo: string,
-    +local_departure: string,
-    +local_arrival: string,
-    +route: Array<ApiRouteItem>,
-    +routes: Array<Array<string>>,
-  |}>,
-|};
+import type {
+  ItinerariesSearchParametersType,
+  ApiResponseType,
+  ItinerariesType,
+} from './ItinerariesloaderTypes';
 
 const dateFormat = 'DD/MM/YYYY';
 
 const parseDate = (date: Date) =>
   DateFNS.format(DateFNS.parse(date), dateFormat);
 
-export const parseParameters = (input: ItinerariesSearchParameters) => {
+export const parseParameters = (input: ItinerariesSearchParametersType) => {
   const params = {
     flyFrom: input.travelFrom,
     dateFrom: parseDate(input.dateFrom),
@@ -174,9 +43,9 @@ export const parseParameters = (input: ItinerariesSearchParameters) => {
 };
 
 const fetchItineraries = async (
-  parameters: $ReadOnlyArray<ItinerariesSearchParameters>,
+  parameters: $ReadOnlyArray<ItinerariesSearchParametersType>,
 ) => {
-  const results: $ReadOnlyArray<ApiResponse> = await Promise.all(
+  const results: $ReadOnlyArray<ApiResponseType> = await Promise.all(
     parameters.map(params => {
       return fetch(`/v2/search?${qs.stringify(parseParameters(params))}`);
     }),
@@ -186,7 +55,7 @@ const fetchItineraries = async (
   });
 };
 
-const sanitizeItineraries = (response: ApiResponse): Itineraries[] => {
+const sanitizeItineraries = (response: ApiResponseType): ItinerariesType[] => {
   const itineraries = response.data;
 
   const newItinerariesStructure = itineraryMock;
@@ -228,8 +97,8 @@ const sanitizeItineraries = (response: ApiResponse): Itineraries[] => {
 export default () =>
   new OptimisticDataloader(
     async (
-      keys: $ReadOnlyArray<ItinerariesSearchParameters>,
-    ): Promise<Array<Itineraries[] | Error>> => fetchItineraries(keys),
+      keys: $ReadOnlyArray<ItinerariesSearchParametersType>,
+    ): Promise<Array<ItinerariesType[] | Error>> => fetchItineraries(keys),
     {
       cacheKeyFn: stringify,
     },
