+++ apps/graphql/src/services/GraphQLContext.js
@@ -2,17 +2,21 @@
 
 import DataLoader from 'dataloader';
 
-import createItinerariesLoader, {
-  type ItinerariesSearchParameters,
-  type Itineraries,
-} from '../dataloaders/Itinerariesloader';
+import createItinerariesLoader from '../dataloaders/Itinerariesloader';
+import {
+  type ItinerariesSearchParametersType,
+  type ItinerariesType,
+} from '../dataloaders/ItinerariesloaderTypes';
 import createLocationLoader, {
   type Locations,
 } from '../dataloaders/Locationsloader';
 
 export type GraphqlContextType = {|
   +dataLoader: {|
-    +itineraries: DataLoader<ItinerariesSearchParameters, Itineraries[]>,
+    +itineraries: DataLoader<
+      ItinerariesSearchParametersType,
+      ItinerariesType[],
+    >,
     +locations: DataLoader<{ term: string }, Locations>,
   |},
 |};
